whisper-backend/requirements.txt
fastapi
uvicorn
openai-whisper
pydub
python-multipart